Stop supporting kernel filesystem in userspace
This drops LCFS_FORMAT_COMPOSEFS from both the write and and the
mount APIs, as well as the tools and tests.
