+++ libcomposefs/Makefile-lib.am
@@ -12,7 +12,6 @@ libcomposefs_la_SOURCES = \
                         $(COMPOSEFSDIR)/lcfs-fsverity.c \
                         $(COMPOSEFSDIR)/lcfs-fsverity.h \
                         $(COMPOSEFSDIR)/lcfs-writer-erofs.c \
-                        $(COMPOSEFSDIR)/lcfs-writer-cfs.c \
                         $(COMPOSEFSDIR)/lcfs-writer.c \
                         $(COMPOSEFSDIR)/lcfs-writer.h \
                         $(COMPOSEFSDIR)/lcfs-utils.h \
