--- tools/test-assets/config-with-hard-link.json.gz.sha256_composefs
@@ -1,1 +1,0 @@
-23a1490815128a431d82143566e6c52858660a42afc4fa1b152f301bed243396
