--- tools/test-assets/config.json.gz.sha256_composefs
@@ -1,1 +1,0 @@
-ac6faf7d04900532563675e71642a01c089c6b6f57828792fc7a2511e7594621
