--- tools/test-assets/cs9-x86_64-developer.json.gz.sha256_composefs
@@ -1,1 +1,0 @@
-0ece6f3b68e0ac258dca9afe66dd46027b954b0cbea5d691ed7eb7c91f070ec3
