--- tools/test-assets/cs9-x86_64-minimal.json.gz.sha256_composefs
@@ -1,1 +1,0 @@
-4d3600a0c6218ea0f716f6b4d86d85310618e03ca6ff035f89a98f780c494d5d
