--- tools/test-assets/f36-x86_64-silverblue.json.gz.sha256_composefs
@@ -1,1 +1,0 @@
-a9a74619e4402b76cbf1a2ce3a0464f52e375b436ad0aab02e24989df0531c62
